refactor: simplify test module structure
Just moves code around, having `mod tests2` is a bit ridiculous.
